perf: Change secret
